Split controllers and move in respective folder
Split directives and services in separate files and moved to respective namespaces and folders
