+++ src/js/angular/import/app.js
@@ -1,7 +1,8 @@
 import 'angular/core/services';
 import 'angular/core/directives';
-import 'angular/import/controllers';
-import 'angular/import/directives';
+import 'angular/import/controllers/import-view.controller';
+import 'angular/import/directives/validate-uri.directive';
+import 'angular/import/directives/import-progress-indicator.directive';
 import 'angular/import/directives/import-resource-tree.directive';
 import 'angular/core/services/repositories.service';
 import 'ng-file-upload/dist/ng-file-upload.min';
@@ -15,8 +16,9 @@ const modules = [
     'graphdb.framework.impex.import.directives',
     'graphdb.framework.impex.import.controllers',
     'graphdb.framework.core.services.repositories',
-    'graphdb.framework.core.directives',
-    'graphdb.framework.import.import-resource-tree'
+    'graphdb.framework.import.directives.validate-uri',
+    'graphdb.framework.import.directives.import-progress-indicator',
+    'graphdb.framework.import.directives.import-resource-tree'
 ];
 
 angular.module('graphdb.framework.impex.import', modules);
