+++ src/pages/import.html
@@ -12,7 +12,7 @@
 
     <div core-errors write ontop></div>
 
-    <div ng-controller="TabCtrl" ng-show="!isRestricted && canWriteActiveRepo()">
+    <div ng-controller="TabController" ng-show="!isRestricted && canWriteActiveRepo()">
 
         <ul class="nav nav-tabs mb-2">
             <li class="nav-item" id="wb-import-tabUpload">
